Rename middleware.ts to /middleware/auth.ts
